index: Use addScriptTag
When you use addScriptTag instead of evaluate the exceptions
thrown get shown in the devtools in chrome.

When you use evaluate the exception is invisible in devtools
and is instead thrown in `page.evaluate()` in the `tape-puppet`
itself.

When trying to debug what's going on in chrome it's best to have
all the debugging happen with the devtools.